Data.Conduit.Binary

Description

Functions for interacting with bytes.

Files and Handles

Note that most of these functions live in the MonadResource monad to ensure resource finalization even in the presence of exceptions. In order to run such code, you will need to use runResourceT.

Sources

sourceFile :: MonadResource m => FilePath -> GSource m ByteString

Stream the contents of a file as binary data.

Since 0.3.0

sourceHandle :: MonadIO m => Handle -> GSource m ByteString

Stream the contents of a Handle as binary data. Note that this function will not automatically close the Handle when processing completes, since it did not acquire the Handle in the first place.

Since 0.3.0

sourceIOHandle :: MonadResource m => IO Handle -> GSource m ByteString

An alternative to sourceHandle. Instead of taking a pre-opened Handle, it takes an action that opens a Handle (in read mode), so that it can open it only when needed and closed it as soon as possible.

Since 0.3.0

sourceFileRange

Arguments

:: MonadResource m

=> FilePath

-> Maybe Integer
Offset

-> Maybe Integer
Maximum count

-> GSource m ByteString

Stream the contents of a file as binary data, starting from a certain offset and only consuming up to a certain number of bytes.

Since 0.3.0

Sinks

sinkFile :: MonadResource m => FilePath -> GInfSink ByteString m

Stream all incoming data to the given file.

Since 0.3.0

sinkHandle :: MonadIO m => Handle -> GInfSink ByteString m

Stream all incoming data to the given Handle. Note that this function will not automatically close the Handle when processing completes.

Since 0.3.0

sinkIOHandle :: MonadResource m => IO Handle -> GInfSink ByteString m

An alternative to sinkHandle. Instead of taking a pre-opened Handle, it takes an action that opens a Handle (in write mode), so that it can open it only when needed and close it as soon as possible.

Since 0.3.0

Conduits

conduitFile :: MonadResource m => FilePath -> GInfConduit ByteString m ByteString

Stream the contents of the input to a file, and also send it along the pipeline. Similar in concept to the Unix command tee.

Since 0.3.0

Utilities

Sources

sourceLbs :: Monad m => ByteString -> GSource m ByteString

Stream the chunks from a lazy bytestring.

Since 0.5.0

Sinks

head :: Monad m => GLSink ByteString m (Maybe Word8)

Return the next byte from the stream, if available.

Since 0.3.0

dropWhile :: Monad m => (Word8 -> Bool) -> GLSink ByteString m ()

Ignore all bytes while the predicate returns True.

Since 0.3.0

take :: Monad m => Int -> GLSink ByteString m ByteString

Take the given number of bytes, if available.

Since 0.3.0

drop :: Monad m => Int -> GLSink ByteString m ()

Drop up to the given number of bytes.

Since 0.5.0

Conduits

isolate :: Monad m => Int -> GLConduit ByteString m ByteString

Ensure that only up to the given number of bytes are consume by the inner sink. Note that this does not ensure that all of those bytes are in fact consumed.

Since 0.3.0

takeWhile :: Monad m => (Word8 -> Bool) -> GLConduit ByteString m ByteString

Return all bytes while the predicate returns True.

Since 0.3.0

lines :: Monad m => GInfConduit ByteString m ByteString

Split the input bytes into lines. In other words, split on the LF byte (10), and strip it from the output.

Since 0.3.0