Data.Conduit.Lazy

Description

Use lazy I/O for consuming the contents of a source. Warning: All normal warnings of lazy I/O apply. In particular, if you are using this with a ResourceT transformer, you must force the list to be evaluated before exiting the ResourceT.

Documentation

lazyConsume :: (MonadBaseControl IO m, MonadActive m) => Pipe l i a () m r -> m [a]

Use lazy I/O to consume all elements from a Source.

This function relies on monadActive to determine if the underlying monadic state has been closed.

Since 0.3.0