Data.Conduit.Text

Description

Handle streams of text.

Parts of this code were taken from enumerator and adapted for conduits.

Text codecs

data Codec

A specific character encoding.

Since 0.3.0

encode :: MonadThrow m => Codec -> GInfConduit Text m ByteString

Convert text into bytes, using the provided codec. If the codec is not capable of representing an input character, an exception will be thrown.

Since 0.3.0

decode :: MonadThrow m => Codec -> GInfConduit ByteString m Text

Convert bytes into text, using the provided codec. If the codec is not capable of decoding an input byte sequence, an exception will be thrown.

Since 0.3.0

utf8 :: Codec

Since 0.3.0

utf16_le :: Codec

Since 0.3.0

utf16_be :: Codec

Since 0.3.0

utf32_le :: Codec

Since 0.3.0

utf32_be :: Codec

Since 0.3.0

ascii :: Codec

Since 0.3.0

iso8859_1 :: Codec

Since 0.3.0

lines :: Monad m => GInfConduit Text m Text

Emit each line separately

Since 0.4.1